Rename getrepository to get package
