+++ server/application/src/main/kotlin/io/vitalir/kotlinhub/server/app/feature/repository/routes/RepositoryRoutes.kt
@@ -3,7 +3,7 @@ package io.vitalir.kotlinhub.server.app.feature.repository.routes
 import io.ktor.server.routing.*
 import io.vitalir.kotlinhub.server.app.common.routes.jwtAuth
 import io.vitalir.kotlinhub.server.app.feature.repository.routes.create.createRepositoryRoute
-import io.vitalir.kotlinhub.server.app.feature.repository.routes.getrepository.getRepositoryRoute
+import io.vitalir.kotlinhub.server.app.feature.repository.routes.get.getRepositoryRoute
 import io.vitalir.kotlinhub.server.app.feature.repository.routes.remove.removeRepositoryForCurrentUserRoute
 import io.vitalir.kotlinhub.server.app.feature.repository.routes.update.updateRepositoryForCurrentUserRoute
 import io.vitalir.kotlinhub.server.app.infrastructure.di.AppGraph
