+++ server/application/src/main/kotlin/io/vitalir/kotlinhub/server/app/feature/repository/routes/get/GetRepositoryResponse.kt
@@ -1,4 +1,4 @@
-package io.vitalir.kotlinhub.server.app.feature.repository.routes.getrepository
+package io.vitalir.kotlinhub.server.app.feature.repository.routes.get
 
 import io.vitalir.kotlinhub.server.app.feature.repository.routes.ApiRepository
 import kotlinx.serialization.Serializable
